Port phoneME (CLDC/MIDP) to run natively on PS Vita
Adds a full vita_arm build target across CLDC, MIDP, and PCSL: native
SDL2 input, the Chameleon LCDUI rendering pipeline, a com.nokia.mid.ui
stub subsystem, and a from-scratch JSR184 (M3G) 3D engine with its own
binary .m3g loader and software rasterizer. Fixes several native-layer
bugs found along the way (input starvation from a never-wired VM timer
tick, Chameleon swallowing softkeys outside full-screen mode, and a
broken native trig implementation replaced system-wide in
java.lang.Math). Built on top of j2me-preservation/phoneME-GP2X-SDL.

+++ phoneme_feature/cldc/src/vm/cpu/arm/Assembler_arm.hpp
@@ -771,7 +771,7 @@ friend struct OpcodeInfo;
 #endif
 
   void pld(Address2 addr) {
-    GUARANTEE((addr & mode_flags) == offset, "must be offset");
+    GUARANTEE(((int)addr & (int)mode_flags) == (int)offset, "must be offset");
     emit(0xF450F000 | addr);
   }
